Added default to Synthetic calibrator and new constructor. Moved MarketDataRatesProvider to pricer.
