+++ modules/function/src/main/java/com/opengamma/strata/function/calculation/swaption/SwaptionMeasureCalculations.java
@@ -9,9 +9,9 @@ import com.opengamma.strata.basics.currency.CurrencyAmount;
 import com.opengamma.strata.basics.market.MarketData;
 import com.opengamma.strata.calc.marketdata.CalculationMarketData;
 import com.opengamma.strata.calc.runner.function.result.CurrencyValuesArray;
-import com.opengamma.strata.function.marketdata.MarketDataRatesProvider;
 import com.opengamma.strata.market.key.SwaptionVolatilitiesKey;
 import com.opengamma.strata.market.view.SwaptionVolatilities;
+import com.opengamma.strata.pricer.rate.MarketDataRatesProvider;
 import com.opengamma.strata.pricer.rate.RatesProvider;
 import com.opengamma.strata.pricer.swaption.VolatilitySwaptionCashParYieldProductPricer;
 import com.opengamma.strata.pricer.swaption.VolatilitySwaptionPhysicalProductPricer;
